DM-84 - produce prototype of more efficient compression calculation in DuckDB
This aligns with the proposal in DM-86.

NOTE: if we go this direction, then we need to make sure we have access to all of the sizes for the various
data types provided by DuckDB, as noted by FIXME comments in the source. This prototype only has a handful,
so we can test with the `nyc311` dataset.