Add Salesforce Connector
Add Salesforce connector supporting the FileIterator interface using the
Bulk API with CSV jobs.

Each source requires a SOQL Query and SObject name.

Credentials can be provided for the source or the connector.
Username/password and JWT authentication is supported.

Each bulk job batch result is retrieved as a single CSV file for rill to
ingest.