Removed Trial information
Indeed, I agree with @AndrewRTsao and will go ahead and remove the trial section for now!

I think its worth to have at least the updated pages pushed to main and then we can add the trial information when ready. I will remove the contents with @ericokuma when im State-side to ensure that we're capturing all the correct information.